convertSBtoNormal {EcoCluster}	R Documentation

Convert the vector v into the vector theta/phi using the stick-breaking formula

Description

This function converts the input vector v into another vector of probabilities that sum to one based on the stick-breaking equation

Usage

convertSBtoNormal(v)

Arguments

v
vector of probabilities (these do not have to sum to 1)

Value

this function returns a vector res with probabilities that sum to one

[Package EcoCluster version 1.0 Index]