mixture.gibbs.main.func {EcoCluster}	R Documentation

Main function of the Mixture Model

Description

Runs the Gibbs sampler and returns samples from the posterior distribution

Usage

mixture.gibbs.main.func(dat, ngroup, ngibbs, burnin)

Arguments

dat
this matrix has L rows (locations) and S columns (species) and contains the presence-absence data

ngroup
maximum number of location groups (K)

ngibbs
number of Gibbs sampler iterations

burnin
number of iterations to discard as burn-in

Value

this function returns a list containing several matrices. These matrices have ngibbs-burnin rows and contain samples from the posterior distribution for:

phi: probability of observing each species in each group

theta: probability of each location group

logl: log-likelihood

z: cluster assignment of each location

gamma: TSB prior parameter

[Package EcoCluster version 1.0 Index]