ncs {EcoCluster}	R Documentation

Summarize the data for mixture model

Description

This function summarizes the data by calculating the number of locations for each location group and species for which dat(i,j)=1 (stored in res1) and dat(i,j)=0 (stored in res0)

Usage

ncs(dat, z, nspp, nloc, ngroup)

Arguments

dat
matrix with L rows (e.g., locations) and S columns (e.g., species), containing the presence-absence data

z
vector of size L with cluster assignment for each location

nspp
number of species (S)

nloc
number of locations (L)

ngroup
maximum number of groups (K)

Value

this function returns a list containing two matrices of size K x S: res1 and res0

[Package EcoCluster version 1.0 Index]