sample.betas {EcoCluster}	R Documentation

Sample betas

Description

This function generates samples of betas, the regression slope parameters

Usage

sample.betas(ngroups, cs, nparam, xtx, t.xmat, alpha, nloc, nspp, omega)

Arguments

ngroups
maximum number of species groups (KS)

cs
vector of size S with group assignment for each species

nparam
overall number of covariates (P)

xtx
equal to t(X)*X

t.xmat
equal to tranpose of the design matrix t(X)

alpha
vector of size S with intercept for each species

nloc
overall number of locations (L)

nspp
overall number of species (S)

omega
a matrix L x S containing the underlying latent variables omega

Value

this function returns a P x KS matrix with regression slope parameters betas

[Package EcoCluster version 1.0 Index]