sample.omega {EcoCluster}	R Documentation

Sample omega

Description

This function generates samples of omega (the underlying latent continuous variable)

Usage

sample.omega(y, nspp, nloc, xmat, betas, cs, alpha)

Arguments

y
L x S matrix, containing the presence-absence data

nspp
overall number of species S

nloc
overall number of locations L

xmat
L x P matrix containing the covariates (columns) for each location (rows). Notice that this matrix does not contain a column of 1's for the intercept

betas
P x KS matrix containing slope parameters for each covariate (row) for each group (column)

cs
group assignment for each species

alpha
intercept for each species

n
number of samples to generate

Value

this function returns a L x S matrix containing the underlying latent variables omega

[Package EcoCluster version 1.0 Index]