sample.psi {EcoCluster}	R Documentation

Sample psi parameters

Description

Sample the presence probability for each location and species group (psi) from it's full conditional distribution

Usage

sample.psi(z, w, dat, ngroup.loc, ngroup.spp)

Arguments

z
vector of length L, storing the current membership of each location

w
vector of length S, storing the current membership of each species

dat
L x S matrix containing the presence-absence data

ngroup.loc
maximum number of location groups (KL)

ngroup.spp
maximum number of species groups (KS)

Value

a KL x KS matrix of psi parameters

[Package EcoCluster version 1.0 Index]