tnorm {EcoCluster}	R Documentation

Sample from a truncated normal distribution

Description

This function generates n samples from a truncated normal distribution using the CDF of the normal distribution

Usage

tnorm(n, lo, hi, mu, sig)

Arguments

n
number of samples to generate

lo
lower truncation point. If there is no lower truncation, lo should be set to -Inf

hi
upper truncation point. If there is no upper truncation, hi should be set to Inf

mu
mode of the truncated normal distribution

sig
dispersion parameter for truncated normal

Value

this function returns a vector of size n

[Package EcoCluster version 1.0 Index]