update.z {EcoCluster}	R Documentation

Samples z

Description

This function samples the cluster assignment for each location (z)

Usage

## S3 method for class 'z'
update(dat, one.minus.dat, phi, theta, ngroup, nloc, nspp, z)

Arguments

dat
matrix with L rows (e.g., locations) and S columns (e.g., species), containing the presence-absence data

one.minus.dat
matrix with L rows (e.g., locations) and S columns (e.g., species), calculated as 1-dat

phi
K x S matrix with the probability of observing each species in each group

theta
vector of length K with the probability of each location group

ngroup
maximum number of location groups (K)

nloc
number of locations (L)

nspp
number of species (S)

z
vector of size L containing the current cluster assignment for each location

Value

this function returns a vector of size L with the cluster assignment of each location

[Package EcoCluster version 1.0 Index]